TR-12984 Restructure the repository to have fewer packages
