--- lib/cmd/command.go
@@ -1,15 +1,0 @@
-package cmd
-
-import (
-	"github.com/Travix-International/Travix.Core.Adk/lib/context"
-)
-
-type Command struct {
-	Verbose       bool
-	LocalFrontend bool
-	TargetEnv     string
-}
-
-type Registrable interface {
-	Register(context.Context)
-}
